common素材を切り出し
