+++ app/src/main/java/com/kabos/topicker/ui/common/ScalableButton.kt
@@ -1,0 +1,79 @@
+package com.kabos.topicker.ui.common
+
+import androidx.compose.animation.core.Animatable
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.scale
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.kabos.topicker.ui.theme.TopickerTheme
+import kotlinx.coroutines.launch
+
+
+@Composable
+fun ScalableButton(
+    onClick: () -> Unit,
+    isEnabled: Boolean = false,
+    text: String,
+    activeColor: Color = Color(0xFF35898F),
+    activeTextColor: Color = Color.White,
+    animationDuration: Int = 100,
+    scaleDown: Float = 0.9f,
+    modifier: Modifier = Modifier
+) {
+    val interactionSource = MutableInteractionSource()
+    val coroutineScope = rememberCoroutineScope()
+    val scale = remember { Animatable(1f) }
+
+    Box(
+        modifier = modifier
+            .scale(scale = scale.value)
+            .background(
+                color = if (isEnabled) activeColor else Color.LightGray,
+                shape = RoundedCornerShape(size = 12f)
+            )
+            .clickable(interactionSource = interactionSource, indication = null) {
+                onClick()
+                coroutineScope.launch {
+                    scale.animateTo(
+                        scaleDown,
+                        animationSpec = tween(animationDuration),
+                    )
+                    scale.animateTo(
+                        1f,
+                        animationSpec = tween(animationDuration),
+                    )
+                }
+            }
+    ) {
+        Text(
+            text = text,
+            modifier = Modifier.padding(horizontal = 24.dp, vertical = 12.dp),
+            color = if (isEnabled) activeTextColor else Color.White,
+            fontWeight = FontWeight.Medium
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun PreviewScalableButton() {
+    TopickerTheme {
+        ScalableButton(
+            onClick = {},
+            text = "Button"
+        )
+    }
+}
